# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FormalAP1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/results": {
            "get": {
                "tags": [
                    "Results"
                ],
                "description": "Return all result with their informations.",
                "operationId": "getAllresult",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/results/{result}": {
            "get": {
                "tags": [
                    "Results"
                ],
                "description": "Return result by ID.",
                "operationId": "getAllresult",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Results"
                ],
                "description": "Update result by Id.",
                "operationId": "UpdateResult",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Results"
                ],
                "description": "Update result by Id.",
                "operationId": "UpdateResult",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/results/": {
            "post": {
                "tags": [
                    "Results"
                ],
                "description": "Create new result.",
                "operationId": "Createresult",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        }
    },
    "security": [
        {
            "constants": {
                "L5_SWAGGER_CONST_HOST": "http://127.0.0.1:8000/api"
            }
        }
    ]
}